requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
fastapi==0.88.0
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.31
protobuf==3.20.1
pydantic==1.10.2
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
numpy==1.24.2